virtio: Rename block and console
