+++ src/Makefile
@@ -65,8 +65,8 @@ OS_OBJECTS = kernel/kernel_start.o kernel/syscalls.o kernel/vga.o \
 		crt/c_abi.o crt/string.o crt/quick_exit.o crt/cxx_abi.o  crt/mman.o \
 		util/memstream.o \
 		hw/pit.o hw/pic.o hw/pci_device.o hw/cpu_freq_sampling.o \
-		virtio/virtio.o virtio/virtionet.o virtio/virtio_blk.o \
-		virtio/virtio_con.o virtio/virtio_queue.o \
+		virtio/virtio.o virtio/virtio_queue.o virtio/virtionet.o \
+        virtio/block.o virtio/console.o \
 		net/ethernet.o net/inet_common.o net/arp.o net/ip4.o \
 		net/tcp.o net/tcp_connection.o net/tcp_connection_states.o \
 		net/ip4/icmpv4.o net/ip4/udp.o net/ip4/udp_socket.o \
